+++ src/frontend/src/layouts/components/Logo.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>
+    <img src="@/assets/images/logo/logo.png" alt="logo" height="35" />
+  </div>
+</template>